Revert "Merge pull request #19 from agchang94/mikeMerge"
This reverts commit a844b031efdf9d0c7c4ac188b021225d0b093d83, reversing
changes made to 0af35d2e9599a57b3f6b3d1a8157ebbdfb4b2c4f.

+++ seeker/src/main/resources/schema.sql
@@ -10,14 +10,12 @@ DROP TABLE IF EXISTS STUDENT_RESUME;
 
 CREATE SEQUENCE hibernate_sequence START WITH 100 INCREMENT BY 1;
 
-/*
 CREATE TABLE dummy_jobs (
     jobID           INT AUTO_INCREMENT PRIMARY KEY,
     jobTitle        VARCHAR(50),
     jobInfo         VARCHAR(250),
     PRIMARY KEY(jobID)
 );
-*/
 
 CREATE TABLE companies (
     companyID       INT AUTO_INCREMENT PRIMARY KEY,    
@@ -30,7 +28,7 @@ CREATE TABLE companies (
 CREATE TABLE company_jobs (
     id              INT AUTO_INCREMENT PRIMARY KEY,
     company_ID      INT(8),
-    job_ID          BIGINT(19),
+    job_ID          INT(8),
     PRIMARY KEY(id)
 );
